requirements.txt
Scrapy==1.5.0
pandas==0.22.0
google-cloud-translate==1.3.1
betamax-0.8.1
